Updated treepy.py
A new version edited from the previous, using nano on my phone, so it might not look very neat.
This adds the capability to give the directory you want to run the tree in as an argument to the script. I could not access external storage via Termux without using `su`, but the superuser in Termux did not have the python binaries; necessitating this workaround.